add floats suuport

+++ PolygonOperations/PolygonOperations/Source.cpp
@@ -31,8 +31,8 @@ vector<std::string> split(string &text, char sep) {
 
 ///////////////// STRUCTURES ////////////////////
 struct Point {
-	int x;
-	int y;
+	float x;
+	float y;
 };
 
 struct Polygon {
@@ -89,8 +89,8 @@ void parse_file(string file_name) {
 			remove_delim_y.erase(std::remove(remove_delim_y.begin(), remove_delim_y.end(), ')'), remove_delim_y.end());
 
 			Point point;
-			point.x = atoi(remove_delim_x.c_str());
-			point.y = atoi(remove_delim_y.c_str());
+			point.x = atof(remove_delim_x.c_str());
+			point.y = atof(remove_delim_y.c_str());
 
 			polygon_points[index++] = point;
 		}
